+++ plugins/keyboard-navigation/test/navigation_modify_test.mocha.js
@@ -5,7 +5,7 @@
  */
 
 const chai = require('chai');
-const Blockly = require('blockly/node');
+const Blockly = require('blockly');
 const {Navigation} = require('../src/navigation');
 const assert = chai.assert;
 const {testHelpers} = require('@blockly/dev-tools');