+++ plugins/workspace-minimap/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.14](https://github.com/google/blockly-samples/compare/@blockly/workspace-minimap@0.1.13...@blockly/workspace-minimap@0.1.14) (2024-02-08)
+
+**Note:** Version bump only for package @blockly/workspace-minimap
+
+
+
+
+
 ## [0.1.13](https://github.com/google/blockly-samples/compare/@blockly/workspace-minimap@0.1.12...@blockly/workspace-minimap@0.1.13) (2024-01-11)